fix: 도메인 별 Exception class 수정
