+++ src/main/java/com/example/umc10th/domain/mission/exception/MissionException.java
@@ -1,7 +1,10 @@
 package com.example.umc10th.domain.mission.exception;
 
-public class MissionException extends RuntimeException {
-    public MissionException(String message) {
-        super(message);
+import com.example.umc10th.domain.mission.exception.code.MissionErrorCode;
+import com.example.umc10th.global.apiPayload.exception.ProjectException;
+
+public class MissionException extends ProjectException {
+    public MissionException(MissionErrorCode errorCode) {
+        super(errorCode);
     }
 }
