+++ src/main/java/com/example/umc10th/domain/mission/exception/code/MissionErrorCode.java
@@ -1,4 +1,18 @@
 package com.example.umc10th.domain.mission.exception.code;
 
-public enum MissionErrorCode {
+import com.example.umc10th.global.apiPayload.code.BaseErrorCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum MissionErrorCode implements BaseErrorCode {
+
+    MISSION_NOT_FOUND(HttpStatus.NOT_FOUND, "MISSION404_1", "해당 미션을 찾을 수 없습니다."),
+    ;
+
+    private final HttpStatus status;
+    private final String code;
+    private final String message;
 }
