+++ src/main/java/com/example/umc10th/domain/review/controller/ReviewController.java
@@ -1,7 +1,7 @@
 package com.example.umc10th.domain.review.controller;
 
 
-import com.example.umc10th.domain.review.code.ReviewSuccessCode;
+import com.example.umc10th.domain.review.exception.code.ReviewSuccessCode;
 import com.example.umc10th.domain.review.dto.ReviewReqDTO;
 import com.example.umc10th.domain.review.dto.ReviewResDTO;
 import com.example.umc10th.domain.review.service.ReviewService;
