+++ src/main/java/com/example/umc10th/domain/review/exception/ReviewException.java
@@ -1,7 +1,10 @@
 package com.example.umc10th.domain.review.exception;
 
-public class ReviewException extends RuntimeException {
-    public ReviewException(String message) {
-        super(message);
+import com.example.umc10th.domain.review.exception.code.ReviewErrorCode;
+import com.example.umc10th.global.apiPayload.exception.ProjectException;
+
+public class ReviewException extends ProjectException {
+    public ReviewException(ReviewErrorCode errorCode) {
+        super(errorCode);
     }
 }
