+++ src/main/java/com/example/umc10th/domain/user/exception/UserException.java
@@ -1,7 +1,10 @@
 package com.example.umc10th.domain.user.exception;
 
-public class UserException extends RuntimeException {
-    public UserException(String message) {
-        super(message);
+import com.example.umc10th.domain.user.exception.code.UserErrorCode;
+import com.example.umc10th.global.apiPayload.exception.ProjectException;
+
+public class UserException extends ProjectException {
+    public UserException(UserErrorCode errorCode) {
+        super(errorCode);
     }
 }
